+++ Avalonia.Lottie/Parser/LottieCompositionParser.cs
@@ -69,8 +69,8 @@ namespace Avalonia.Lottie.Parser
 
             reader.EndObject();
 
-            var scaledWidth = (int) (width );
-            var scaledHeight = (int) (height);
+            var scaledWidth = width;
+            var scaledHeight = height;
             var bounds = new Rect(0, 0, scaledWidth, scaledHeight);
 
             composition.Init(bounds, startFrame, endFrame, frameRate, layers, layerMap, precomps, images, characters,
